requirements.txt
nicegui>=1.3.15
python-dotenv
gunicorn
fastapi
authlib
python-jose
pymongo
python-dotenv
uvicorn
PyJWT